+++ crates/sol-macro/src/expand/event.rs
@@ -1,7 +1,7 @@
 //! [`ItemEvent`] expansion.
 
 use super::{anon_name, expand_tuple_types, expand_type, ExpCtxt};
-use crate::expand::r#type::expand_event_tokenize_func;
+use crate::expand::ty::expand_event_tokenize_func;
 use ast::{EventParameter, ItemEvent, SolIdent};
 use proc_macro2::TokenStream;
 use quote::{quote, quote_spanned};
