+++ crates/sol-types/src/lib.rs
@@ -183,7 +183,7 @@ mod impl_core;
 mod types;
 pub use types::{
     data_type as sol_data, ContractError, Encodable, EventTopic, Panic, PanicKind, Revert,
-    Selectors, SolCall, SolCalls, SolEnum, SolError, SolEvent, SolStruct, SolType, TopicList,
+    Selectors, SolCall, SolEnum, SolError, SolEvent, SolInterface, SolStruct, SolType, TopicList,
 };
 
 mod util;
